Refactor SBIOrderExecutor by splitting responsibilities
