+++ project/modules/trading/sbi/orders/manager/__init__.py
@@ -1,1 +1,5 @@
 from .order_executor import SBIOrderExecutor
+from .order_placer import SBIOrderPlacer
+from .order_canceller import SBIOrderCanceller
+from .order_inquiry import SBIOrderInquiry
+from .position_manager import SBIPositionManager
